Creating Reservation controller and its methods

+++ BACKEND-APPLICATION/src/main/java/com/application/controllers/UserController.java
@@ -95,7 +95,7 @@ public class UserController extends ExceptionHandlingController {
 
     @PutMapping("/user/update/user-role/{id}")
    // @PreAuthorize("hasAnyAuthority('user:update')")
-    public ResponseEntity<HttpResponse> updateUserRole(@PathVariable("id") Long id,   @RequestParam("role") String role) throws IOException {
+    public ResponseEntity<HttpResponse> updateUserRole(@PathVariable("id") Integer id,   @RequestParam("role") String role) throws IOException {
         userServiceBean.updateUserRole(id, role);
         return response(OK, USER_ROLE_UPDATED_SUCCESSFULLY);
     }
